Add product rating when user added review

+++ database/migrations/2020_07_21_084200_create_products_table.php
@@ -21,7 +21,7 @@ class CreateProductsTable extends Migration
             $table->text('product_spec');
             $table->integer('product_price');
             $table->string('product_image')->nullable();
-            $table->integer('product_rating')->nullable();
+            $table->float('product_rating')->nullable();
             $table->timestamps();
         });
     }
